refactor: Code refactored in API and tests completed
